+++ RaytraceGroundUp/Chapter28.7_linux/GlobalTrace.h
@@ -5,17 +5,17 @@
 
 
 class GlobalTrace: public Tracer {
-	public:
+    public:
 
-		GlobalTrace(void);
+        GlobalTrace(void);
 
-		GlobalTrace(World* _worldPtr);
+        GlobalTrace(World* _worldPtr);
 
-		virtual RGBColor
-		trace_ray(const Ray ray, const int depth) const;
+        virtual RGBColor
+        trace_ray(const Ray ray, const int depth) const;
 
-		virtual RGBColor
-		trace_ray(const Ray ray, double& tmin, const int depth) const;
+        virtual RGBColor
+        trace_ray(const Ray ray, float& tmin, const int depth) const;
 };
 
 #endif // GLOBALTRACE_H
